+++ Cmonk.h
@@ -10,7 +10,7 @@ public:
 	Cmonk();
 	~Cmonk();
 
-	virtual HRESULT init(POINT position, float HP, float damage, float exp,float trace);
+	virtual HRESULT init(POINT position, float HP, float damage, int exp,float trace);
 	virtual void release();
 	virtual void update();
 	virtual void render();
